Move layout folder to components
