+++ src/components/table/table.styles.ts
