+++ src/containers/domains-page/config/domains-page-query-params.config.ts
@@ -1,5 +1,5 @@
 import { PageQueryParam } from '@/hooks/use-page-query-params/use-page-query-params.types';
-import { SortingOrder } from '@/layout/table/table.types';
+import { SortingOrder } from '@/components/table/table.types';
 
 const domainPageQueryParamsConfig: [
   PageQueryParam<'searchText', string>,
