+++ src/containers/domains-page/config/domains-table-columns.config.ts
@@ -1,4 +1,4 @@
-import { TableColumn } from '@/layout/table/table.types';
+import { TableColumn } from '@/components/table/table.types';
 import { DomainData } from '../domains-page.types';
 import DomainsTableClusterCell from '../domains-table-cluster-cell/domains-table-cluster-cell';
 import DomainsTableDomainNameCell from '../domains-table-domain-name-cell/domains-table-domain-name-cell';
