+++ src/containers/domains-page/domains-table/domains-table.tsx
@@ -2,7 +2,7 @@
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { Cell, Grid } from 'baseui/layout-grid';
 
-import Table from '@/layout/table/table';
+import Table from '@/components/table/table';
 import usePageQueryParams from '@/hooks/use-page-query-params/use-page-query-params';
 import sortBy, { SortByReturnValue, toggleSortOrder } from '@/utils/sort-by';
 import useStyletronClasses from '@/hooks/use-styletron-classes';
@@ -12,7 +12,7 @@ import domainPageQueryParamsConfig from '../config/domains-page-query-params.con
 import { domainTableColumns } from '../config/domains-table-columns.config';
 
 import type { DomainData } from '../domains-page.types';
-import { SortingOrder } from '@/layout/table/table.types';
+import { SortingOrder } from '@/components/table/table.types';
 import { Props } from './domains-table.types';
 import { cssStyles } from './domains-table.styles';
 import { useInView } from 'react-intersection-observer';
