+++ src/containers/domains-page/domains-table/domains-table.types.ts
@@ -1,4 +1,4 @@
-import type { TableColumn } from '@/layout/table/table.types';
+import type { TableColumn } from '@/components/table/table.types';
 import type { DomainData } from '../domains-page.types';
 
 export type DomainTableColumns = Array<TableColumn<DomainData>>;
